+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Unreleased
  
  - (Added) Priority can be set for transitions (using *low*, *high* or any integer in yaml). transitions
    are selected according to their priorities (still following eventless and inner-first/source state semantics).
+ - (Added) A ``sismic.testing`` module containing some primitives to ease unit testing.
  - (Fixed) State *on entry* time (used for ``idle`` and ``after``) is set after the *on entry* 
    action is executed, making the two predicates more accurate when long-running actions are 
    executed when a state is entered.